move auth into new file
